Create Ex3Client.java
Note: I changed some of the parameters to make sense. An unsigned byte was specified to be sent from the server, but since no data type exists like this is java, I used int instead, as using byte would include the negatives. I did, however, use "readUnsignedByte()" to read this data. Similarly, although short contains 16 bits, which my program uses, it also accounts for negatives. Thus, I used int and sent that to the server. 

Hopefully my program outputs correctly. If not, it's because of this.